# swagger.yaml
openapi: 3.0.0
info:
  title: Barangay Census API
  version: 1.0.0
  description: >-
    An API that allows Barangays to Add people and and barangays and calculate
    the population automatically
  termsOfService: http://swagger.io/terms/
  contact:
    email: 20180015014@my.xu.edu.ph
  license:
    name: Node Js Licence
    url: https://github.com/nodejs/node/blob/main/LICENSE
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    apiKey:
      type: apiKey
      name: APIKey
      in: header
...
paths:
  /barangay:
    get:
      summary: Get a list of barangays with specific filters.
      tags:
        - Barangay
      parameters:
        - name: name/exact
          in: query
          description: Filter by exact name.
          schema:
            type: string
        - name: population/exact
          in: query
          description: Filter by exact population.
          schema:
            type: number
        - name: population/lt
          in: query
          description: Filter by population less than.
          schema:
            type: number
        - name: population/gt
          in: query
          description: Filter by population greater than.
          schema:
            type: number
      security:
        - apiKey: []
      responses:
        '200':
          description: Successful response.
        '400':
          description: No Matching Person.
        '500':
          description: Internal Server error
    post:
      summary: Create a new barangay.
      tags:
        - Barangay
      requestBody:
        description: Barangay data to create
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                population:
                  type: number
      security:
        - apiKey: []
      responses:
        '200':
          description: Barangay created successfully.
        '400':
          description: Clien error.
        '500':
          description: Internal server error.
  /barangay/{barangayID}:
    put:
      summary: Update a barangay.
      tags:
        - Barangay
      parameters:
        - in: path
          name: barangayID
          required: true
          description: ID of the barangay to update.
          schema:
            type: string
      requestBody:
        description: Updated barangay data
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                population:
                  type: number
      security:
        - apiKey: []
      responses:
        '200':
          description: Barangay updated successfully.
        '500':
          description: Internal server error.
    patch:
      summary: Partially update a barangay.
      tags:
        - Barangay
      parameters:
        - in: path
          name: barangayID
          required: true
          description: ID of the barangay to update.
          schema:
            type: string
      requestBody:
        description: Partially updated barangay data
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                population:
                  type: number
      security:
        - apiKey: []
      responses:
        '200':
          description: Barangay updated successfully.
        '500':
          description: Internal server error.
    delete:
      summary: Delete a barangay.
      tags:
        - Barangay
      parameters:
        - in: path
          name: barangayID
          required: true
          description: ID of the barangay to delete.
          schema:
            type: string
      security:
        - apiKey: []
      responses:
        '200':
          description: Barangay deleted successfully.
        '404':
          description: Barangay not found.
        '500':
          description: Internal server error.
...